warehouse function

+++ WMS.DataTransferObject/Dtos/LocationDto.cs
@@ -29,17 +29,17 @@ namespace WMS.DataTransferObject.Dtos
         /// <summary>
         /// 获取或设置储位的长度(m)
         /// </summary>
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         /// <summary>
         /// 获取或设置储位的宽度(m)
         /// </summary>
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         /// <summary>
         /// 获取或设置储位的高度(m)
         /// </summary>
-        public int Height { get; set; }
+        public double Height { get; set; }
 
         /// <summary>
         /// 获取或设置货架 Id
