+++ WMS.Domain/Models/Warehouses/Location.cs
@@ -27,17 +27,17 @@ namespace WMS.Domain.Models.Warehouses
         /// <summary>
         /// 获取或设置储位的长度(m)
         /// </summary>
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         /// <summary>
         /// 获取或设置储位的宽度(m)
         /// </summary>
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         /// <summary>
         /// 获取或设置储位的高度(m)
         /// </summary>
-        public int Height { get; set; }
+        public double Height { get; set; }
 
         /// <summary>
         /// 获取货架 Id
