+++ WMS.Domain/Models/Warehouses/Zone.cs
@@ -34,17 +34,17 @@ namespace WMS.Domain.Models.Warehouses
         /// <summary>
         /// 获取或设置区域的长度(m)
         /// </summary>
-        public int Length { get; set; }
+        public double Length { get; set; }
 
         /// <summary>
         /// 获取或设置区域的宽度(m)
         /// </summary>
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         /// <summary>
         /// 获取或设置区域的高度(m)
         /// </summary>
-        public int Height { get; set; }
+        public double Height { get; set; }
 
         /// <summary>
         /// 获取仓库 Id
